backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-openai
websockets
langchain-google-genai
python-dotenv
numpy
pandas
matplotlib